Separate test files, pass internal names
