+++ lib/markdown.js
@@ -114,4 +114,11 @@ const parseMarkdown = (src) => {
   return scenario;
 };
 
-module.exports = { parseMarkdown };
+module.exports = {
+  Action,
+  Scenario,
+  Step,
+  parseLine,
+  parseStep,
+  parseMarkdown,
+};
