+++ lib/metabot.js
@@ -281,4 +281,4 @@ class Metabot {
   }
 }
 
-module.exports = { Chat, Metabot };
+module.exports = { MENU, Chat, Metabot };
